requirements.txt
# Core dependencies
streamlit==1.30.0
python-dotenv==1.0.0

# OpenAI and AI libraries
openai>=1.30.0
tiktoken>=0.5.0

# Alternative AI provider (FREE)
groq>=0.4.0

# Audio processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.3

# File handling and export
fpdf==1.7.2
python-docx==1.1.0
markdown==3.5.2

# Utilities
requests==2.31.0
tqdm==4.66.1
